+++ driver-core/CHANGELOG.rst
@@ -5,6 +5,8 @@ CHANGELOG
 ------
 
 - [improvement] Make most main objects interface to facilitate testing/mocking (JAVA-195)
+- [api] Renamed shutdown to closeAsync and ShutdownFuture to CloseFuture. Clustering
+  and Session also now implement Closeable (JAVA-247).
 - [bug] Fix potential thread leaks when shutting down Metrics (JAVA-232)
 - [bug] Fix potential NPE in HostConnectionPool (JAVA-231)
 - [bug] Avoid NPE when node is in an unconfigured DC (JAVA-244)